refactor(api): MemberInfo를 index에서 Import
